feat(phone/ui): Make SearchField a reusable component
